use walkOnGoRoutine in any directory instead of root

Here is my local test result:

with original pt:
pt -g Android.mk > /dev/null  194.99s user 15.30s system 202% cpu 1:43.99 total

with my pt (pt_tw):
pt_tw -g Android.mk > /dev/null  315.52s user 41.92s system 461% cpu 1:17.40 total